parse the version out of the version string using locale agnostic g_ascii_strtod instead of sscanf, which would only pick up the integer component with a locale which used a different radix character.

+++ src/trg-client.h
@@ -129,7 +129,7 @@ TrgClient *trg_client_new(void);
 TrgPrefs *trg_client_get_prefs(TrgClient * tc);
 int trg_client_populate_with_settings(TrgClient * tc);
 void trg_client_set_session(TrgClient * tc, JsonObject * session);
-float trg_client_get_version(TrgClient * tc);
+gdouble trg_client_get_version(TrgClient * tc);
 const gchar *trg_client_get_version_string(TrgClient * tc);
 gint64 trg_client_get_rpc_version(TrgClient * tc);
 gchar *trg_client_get_password(TrgClient * tc);
